+++ nab-data-source/src/test/java/ru/hh/nab/datasource/DataSourceContextUnsafeTest.java
@@ -5,14 +5,16 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.MDC;
-import static ru.hh.nab.datasource.DataSourceContextUnsafe.clearMDC;
-import static ru.hh.nab.datasource.DataSourceContextUnsafe.executeInScope;
-import static ru.hh.nab.datasource.DataSourceContextUnsafe.executeOn;
-import static ru.hh.nab.datasource.DataSourceContextUnsafe.getDataSourceName;
-import static ru.hh.nab.datasource.DataSourceContextUnsafe.setDefaultMDC;
 import static ru.hh.nab.jdbc.common.DataSourceType.MASTER;
 import static ru.hh.nab.jdbc.common.DataSourceType.READONLY;
 import static ru.hh.nab.jdbc.common.DataSourceType.SLOW;
+import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.MDC_KEY;
+import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.clearMDC;
+import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.executeInScope;
+import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.executeOn;
+import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.getDataSourceName;
+import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.setDefaultMDC;
+import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.setRequestScopeDataSourceType;
 
 public class DataSourceContextUnsafeTest {
 
@@ -24,51 +26,51 @@ public class DataSourceContextUnsafeTest {
   @Test
   public void testExecuteOn() {
     assertEquals(MASTER, getDataSourceName());
-    assertEquals(MASTER, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+    assertEquals(MASTER, MDC.get(MDC_KEY));
 
     executeOn(SLOW, false, () -> {
       assertEquals(SLOW, getDataSourceName());
-      assertEquals(SLOW, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+      assertEquals(SLOW, MDC.get(MDC_KEY));
 
       executeOn(READONLY, false, () -> {
         assertEquals(READONLY, getDataSourceName());
-        assertEquals(READONLY, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+        assertEquals(READONLY, MDC.get(MDC_KEY));
         return null;
       });
 
       assertEquals(SLOW, getDataSourceName());
-      assertEquals(SLOW, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+      assertEquals(SLOW, MDC.get(MDC_KEY));
       return null;
     });
 
     assertEquals(MASTER, getDataSourceName());
-    assertEquals(MASTER, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+    assertEquals(MASTER, MDC.get(MDC_KEY));
   }
 
   @Test
   public void testClearMDC() {
-    assertEquals(MASTER, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+    assertEquals(MASTER, MDC.get(MDC_KEY));
 
     clearMDC();
 
-    assertNull(MDC.get(DataSourceContextUnsafe.MDC_KEY));
+    assertNull(MDC.get(MDC_KEY));
   }
 
   @Test
   public void testRequestScopeSetOverrideDisabled() {
-    DataSourceContextUnsafe.setRequestScopeDataSourceType("test");
+    setRequestScopeDataSourceType("test");
     executeInScope("test", () -> executeOn(SLOW, false, () -> {
       assertEquals(SLOW, getDataSourceName());
-      assertEquals(SLOW, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+      assertEquals(SLOW, MDC.get(MDC_KEY));
 
       executeOn(READONLY, false, () -> {
         assertEquals(READONLY, getDataSourceName());
-        assertEquals(READONLY, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+        assertEquals(READONLY, MDC.get(MDC_KEY));
         return null;
       });
 
       assertEquals(SLOW, getDataSourceName());
-      assertEquals(SLOW, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+      assertEquals(SLOW, MDC.get(MDC_KEY));
       return null;
     }));
   }
@@ -78,16 +80,16 @@ public class DataSourceContextUnsafeTest {
     var testKey = "test";
     executeInScope(testKey, () -> executeOn(SLOW, true, () -> {
       assertEquals(testKey, getDataSourceName());
-      assertEquals(testKey, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+      assertEquals(testKey, MDC.get(MDC_KEY));
 
       executeOn(READONLY, true, () -> {
         assertEquals(testKey, getDataSourceName());
-        assertEquals(testKey, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+        assertEquals(testKey, MDC.get(MDC_KEY));
         return null;
       });
 
       assertEquals(testKey, getDataSourceName());
-      assertEquals(testKey, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+      assertEquals(testKey, MDC.get(MDC_KEY));
       return null;
     }));
   }
@@ -96,16 +98,16 @@ public class DataSourceContextUnsafeTest {
   public void testRequestScopeUnsetOverrideEnabled() {
     executeOn(SLOW, true, () -> {
       assertEquals(SLOW, getDataSourceName());
-      assertEquals(SLOW, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+      assertEquals(SLOW, MDC.get(MDC_KEY));
 
       executeOn(READONLY, true, () -> {
         assertEquals(READONLY, getDataSourceName());
-        assertEquals(READONLY, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+        assertEquals(READONLY, MDC.get(MDC_KEY));
         return null;
       });
 
       assertEquals(SLOW, getDataSourceName());
-      assertEquals(SLOW, MDC.get(DataSourceContextUnsafe.MDC_KEY));
+      assertEquals(SLOW, MDC.get(MDC_KEY));
       return null;
     });
   }
