+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/datasource/RoutingDataSource.java
@@ -12,7 +12,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.datasource.DelegatingDataSource;
 import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
 import org.springframework.lang.Nullable;
-import ru.hh.nab.datasource.DataSourceContextUnsafe;
 import static ru.hh.nab.datasource.DataSourceSettings.DATASOURCE_NAME_FORMAT;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
 import ru.hh.nab.jdbc.common.DataSourceType;
@@ -20,6 +19,7 @@ import ru.hh.nab.jdbc.common.datasource.NamedDataSource;
 import ru.hh.nab.jdbc.common.ext.JdbcExtension;
 import ru.hh.nab.jdbc.common.healthcheck.HealthCheck;
 import ru.hh.nab.jdbc.common.healthcheck.HealthCheckDataSource;
+import ru.hh.nab.jdbc.routing.DataSourceContextUnsafe;
 import ru.hh.nab.metrics.Counters;
 import ru.hh.nab.metrics.StatsDSender;
 import ru.hh.nab.metrics.Tag;
