+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/routing/DataSourceRoutingFilter.java
@@ -11,9 +11,9 @@ import java.io.IOException;
 import org.glassfish.jersey.uri.UriComponent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import ru.hh.nab.datasource.DataSourceContextUnsafe;
 import ru.hh.nab.hibernate.transaction.DataSourceContext;
 import ru.hh.nab.jdbc.common.DataSourceType;
+import ru.hh.nab.jdbc.routing.DataSourceContextUnsafe;
 
 public class DataSourceRoutingFilter implements Filter {
   private static final Logger LOG = LoggerFactory.getLogger(DataSourceRoutingFilter.class);
