+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/transaction/DataSourceContextTransactionManager.java
@@ -10,8 +10,8 @@ import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.support.DefaultTransactionDefinition;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isActualTransactionActive;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isSynchronizationActive;
-import ru.hh.nab.datasource.DataSourceContextUnsafe;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
+import ru.hh.nab.jdbc.routing.DataSourceContextUnsafe;
 
 public class DataSourceContextTransactionManager implements PlatformTransactionManager {
