+++ nab-jdbc/pom.xml
@@ -16,6 +16,10 @@
     <dependencies>
         <dependency>
             <groupId>ru.hh.nab</groupId>
+            <artifactId>nab-common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>ru.hh.nab</groupId>
             <artifactId>nab-jdbc-common</artifactId>
         </dependency>
     </dependencies>
